+++ src/attestedclaim/AttestedClaim.spec.ts
@@ -1,5 +1,5 @@
 import Identity from '../identity/Identity'
-import AttesterIdentity from '../identity/AttesterIdentity'
+import AttesterIdentity from '../attesteridentity/AttesterIdentity'
 import AttestedClaim from './AttestedClaim'
 import Attestation from '../attestation/Attestation'
 import CType from '../ctype/CType'
